updating branch

+++ server/schemas/resolvers.js
@@ -4,42 +4,80 @@ const { signToken } = require('../utils/auth');
 
 const resolvers = {
     Query: {
-        user: async (parent, { userId }) => {
-            return User.findOne({ _id: userId });
+        // For dev only
+        users: async () => {
+            return await User.find().populate('applications').populate('adoptions');
+        },
+
+        user: async (parent, args) => {
+            return await User.findById(args.id).populate('applications').populate('adoptions');
         },
 
         animals: async () => {
-            return Animal.find();
+            return await Animal.find().populate('applications').populate('adoption');
         },
 
-        animal: async (parent, { animalId }) => {
-            return Animal.findById(animalId);
+        animal: async (parent, args) => {
+            return await Animal.findById(args.id).populate('applications').populate('adoption');
         },
 
-        application: async (parent, { applicationId }) => {
-            return Application.findById(applicationId)
+        application: async (parent, args) => {
+            return await Application.findById(args.animalId)
+        },
+
+        applications: async (parent, args) => {
+            return await Application.find({ adoptee: args.id })
         }
     },
 
     Mutation: {
-        addUser: async (parent,  { username, email, password }) => {
-            return User.create({ username, email, password });
+        addUser: async (parent, { username, email, password }) => {
+            const user = await User.create({ username, email, password });
+            const token = signToken(user);
+
+            return { token, user };
         },
 
-        removeUser: async (parent, { userId }) => {
-            return User.findOneAndDelete({ _id: userId });
+        removeUser: async (parent, { userId }, context) => {
+            if (context.user) {
+                return await User.findOneAndDelete({ _id: userId });
+            }
+
+            throw new AuthenticationError('Not logged in');
         },
 
         addAnimal: async (parent, { name, age, sex, animalType, breed, familyFriendly }) => {
-            return Animal.create({ name, age, sex, animalType, breed, familyFriendly });
+            return await Animal.create({ name, age, sex, animalType, breed, familyFriendly });
         },
 
         removeAnimal: async (parent, { animalId }) => {
-            return Animal.findOneAndDelete({ _id: animalId });
+            return await Animal.findOneAndDelete({ _id: animalId });
         },
 
-        addApplication: async (parent, { applicant, adoptee }) => {
-            return Application.create({ applicant, adoptee });
+        addApplication: async (parent, { applicant, adoptee }, context) => {
+            if (context.user) {
+                return await Application.create({ applicant, adoptee });
+            }
+
+            throw new AuthenticationError('Not logged in');
+        },
+
+        login: async (parent, { email, password }) => {
+            const user = await User.findOne({ email });
+
+            if (!user) {
+                throw new AuthenticationError('Incorrect credentials');
+            }
+
+            const correctPw = await user.isCorrectPassword(password);
+
+            if (!correctPw) {
+                throw new AuthenticationError('Incorrect credentials');
+            }
+
+            const token = signToken(user);
+
+            return { token, user };
         }
     }
 };
